+++ e2e/uimodel/bandwagon/bandwagon.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 
 	"github.com/gravitational/robotest/e2e/framework"
-	"github.com/gravitational/robotest/e2e/model/ui/defaults"
-	"github.com/gravitational/robotest/e2e/model/ui/utils"
+	"github.com/gravitational/robotest/e2e/uimodel/defaults"
+	"github.com/gravitational/robotest/e2e/uimodel/utils"
 
 	log "github.com/Sirupsen/logrus"
 	. "github.com/onsi/gomega"
