+++ e2e/uimodel/opscenter/opscenter.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	"github.com/gravitational/robotest/e2e/framework"
-	"github.com/gravitational/robotest/e2e/model/ui/defaults"
-	"github.com/gravitational/robotest/e2e/model/ui/utils"
+	"github.com/gravitational/robotest/e2e/uimodel/defaults"
+	"github.com/gravitational/robotest/e2e/uimodel/utils"
 
 	log "github.com/Sirupsen/logrus"
