+++ e2e/uimodel/user/login.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/gravitational/log"
 	"github.com/gravitational/robotest/e2e/framework"
-	"github.com/gravitational/robotest/e2e/model/ui/defaults"
-	"github.com/gravitational/robotest/e2e/model/ui/utils"
+	"github.com/gravitational/robotest/e2e/uimodel/defaults"
+	"github.com/gravitational/robotest/e2e/uimodel/utils"
 	"github.com/gravitational/trace"
 
 	. "github.com/onsi/gomega"
